requirements.txt
affine==2.3.1
asgiref==3.6.0
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
Django==4.1.5
django-clearcache==1.2.1
dnspython==2.2.1
docopt==0.6.2
fonttools==4.38.0
GeoAlchemy2==0.12.5
geopandas==0.12.2
geoserver-rest==2.3.4
greenlet==2.0.1
idna==3.4
Js2Py==0.74
kiwisolver==1.4.4
matplotlib==3.6.2
munch==2.5.0
packaging==22.0
Pillow==9.4.0
pipwin==0.5.2
postgres-helper==0.0.5
psycopg2==2.9.5
Pygments==2.14.0
pyjsparser==2.7.1
pymongo==4.3.3
pyparsing==3.0.9
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
rasterio==1.3.4
requests==2.28.1
seaborn==0.12.2
six==1.16.0
snuggs==1.4.7
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
sqlparse==0.4.3
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
xmltodict==0.13.0
